fix(hubspot): score a double (no trunca decimales) + RE-ACTIVAR sync de resultado
UCSP pidió que los resultados de examen SÍ viajen a HubSpot (reclamo O1):
- Creadas las props score_*/max_score_*/ultima_evaluacion en el portal 9013951
  (no existían → el sync daba 400 INVALID_PROPERTY y los jobs morían en DLQ).
- Proto SyncExamResultRequest.score/max_score: int32 → double (regen buf);
  domain.ExamResult float64 + FloatStr (preserva 1.25/pregunta del Nacional);
  enqueuer payload + worker float; gateway sin cast int32().
- Re-activado el `go p.syncResultToHubspot(...)` en finishAttempt (estaba
  comentado desde 2026-06-17 esperando justo esto).

+++ gateway/internal/proxy/hubspot.go
@@ -98,11 +98,11 @@ func (p *Proxy) sendHubspotOTP(w http.ResponseWriter, r *http.Request) {
 }
 
 type syncHubspotResultRequest struct {
-	DNI             string `json:"dni"`
-	ExamTypeCode    string `json:"exam_type_code"`
-	Score           int32  `json:"score"`
-	MaxScore        int32  `json:"max_score"`
-	AttemptID       string `json:"attempt_id"`
+	DNI             string  `json:"dni"`
+	ExamTypeCode    string  `json:"exam_type_code"`
+	Score           float64 `json:"score"`
+	MaxScore        float64 `json:"max_score"`
+	AttemptID       string  `json:"attempt_id"`
 	SubmittedAt     string `json:"submitted_at"`
 	ContactRecordID string `json:"contact_record_id"`
 }
